requirements.txt
colorama==0.4.6
coverage==7.3.2
greenlet==3.0.0
iniconfig==2.0.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pytest==7.4.2
pytest-cov==4.1.0
SQLAlchemy==2.0.22
types-Pillow==10.1.0.2
typing_extensions==4.8.0
